00_Dataset_gen/requirements.txt
pandas>=1.3
numpy==1.22.4
numba==0.55.2
librosa==0.9.2
scipy>=1.9
soundfile>=0.10
pydub>=0.25.1
torchaudio>=0.10
torch>=1.10
tqdm>=4.62
PyYAML>=6.0
audiomentations>=0.30
matplotlib=3.5.3
psola